+++ src/Toml.hs
@@ -9,11 +9,13 @@ import qualified Toml
 module Toml
     ( module Toml.Bi
     , module Toml.Parser
+    , module Toml.PrefixTree
     , module Toml.Printer
     , module Toml.Type
     ) where
 
 import Toml.Bi
 import Toml.Parser
+import Toml.PrefixTree
 import Toml.Printer
 import Toml.Type